requirements.txt
numpy>=1.19.4
opencv-python==4.5.3.56
tensorflow-gpu==2.5.0
matplotlib>=3.3.3
Jetson.GPIO
pyserial>=3.5
Pillow>=8.4.0
scikit-learn>=0.24.2
pytest>=6.2.5